Backend/KnowledgeAssistant/pyproject.toml
[project]
name = "knowledgeassistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "gitpython>=3.1.44",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "passlib>=1.7.4",
    "python-jose>=3.4.0",
    "sqlalchemy>=2.0.40",
]
